* src/GlueXBremsstrahlungGenerator.cc, src/GlueXBremsstrahlungGenerator.hh [rtj]    - these two sources were reversed, straighten it out.
